// frontend/src/backend.d.ts
export interface ClientInfo {
    contactPerson: string;
    email: string;
    company: string;
    contracts: Array<Contract>;
    address: string;
    phone: string;
    verificationStatus: ClientVerificationStatus;
}
// ...
export interface Contract {
    endDate: bigint;
    year: bigint;
    contractText: string;
    startDate: bigint;
}
// ...
export enum ClientVerificationStatus {
    verified = "verified",
    pending = "pending",
    rejected = "rejected"
}